Show fiat depth as well as BTC depth in the trade page's orderbook.
Make the fiat and BTC depths highlight together when hovered over.

Adjust the amounts auto-filled in the trade form when buying BTC.
Some BTC will be traded at a better price than we ask for, so we will
get more than we ask for.  This means we can ask for less than we want
and still get what we want.